Making a push in tests/all_tests to test every exported function.
Hopefully, once I finish that I will expand to every codepath in every
function, but that will take a while.

Also, sadly the R/bioconductor on my workstation is once again on a different version
from that on travis, so some tests are failing there but working on my
machine.