I picked off a funny corner-case when reading count tables.
Here is a neat (annoying) scenario:  Let us say we read a bunch of count
tables and have header=FALSE by default, but the files do have headers.
What then happens to the generated expressionset?  Well, since 1 row
consists of the words of the headers, the entire matrix turns into a set
of characters.  This of course ends badly when trying to do math on
them.
Therefore, I added logic which does an as.numeric() to each set of read
counts when adding them to the nascent matrix, coercing these wayward
rows into NA.  This is better, but still not optimal.  I therefore added
to that logic a is.na() and drop the corresponding row as each table is
read, before adding it to the growing count matrix.

yay!

There is also a set of changes to improve the set of things for which
we can easily do rpkm.  In addition I added a couple lines to the PCA
plotter which improves how plots look with > 5 batches.  I noticed some
changes to biomart when attempting to get orthologs between species
which require one to use the same requested columns in both tables, so I
dropped an argument and use the one argument for both tables.  Finally,
I did some minor cleanups to the eupathdb downloaders, but those are
insufficient to handle the odd cases where it fails because some
eupathdb providers do not like my formatting list.