a couple minor peptide plot changes to make it a little more robust.
Found a funny and I think unused piece of eupathdb code which should
probably get deleted.  It was intended to automagically pull out the
needed parameters for searching tables of the eupathdb, but I do not
think it ever got completed -- with that in mind I honestly cannot say
how it is that my ortholog tables were getting filled in with that code
in there, it is quite mysterious to me.