continuing to make combined de stuff more robust.
With the addition of EBSeq, I found it semi-necessary to rework portions
of the unwieldy code which combines those data frames into a bigger and
hopefully easier to interpret pile of data.  I want to further simplify
that codepath and split out the combine_de_tables and
extract_significant_genes so that they will properly handle arbitraryily
large numbers of tools/methods.  But for the moment, working is
sufficient.