new roxygen, also cleaned up combine_de_tables()
I have been wanting for quite a long time to make a nicer version of my
function which combined outputs from limma/deseq/edger/etc.  But it just
got to unwieldy to mess with.  Last week I finally stripped it down and
pieced it back together in a way which should allow me to make it
generally better.