I got covr working properly, so I went a bit crazy adding tests.
I have also decided to try to have actual working examples in my
roxygen.  This is a bit weird because many of my functions require some
data to work with and make chains with each other.  I don't really want
to have every example include multiple setup/teardown commands, so I
did a little reading and thought I understood how the examples are
included.  I was wrong, but here is what I learned:

The R files are read in the order they are found in the Collate field of
the DESCRIPTION.  In some of my reading, I interpreted something as
saying the files are in lexical order in the R directory; so I renamed
them so that they would eventually match the numbers of the tests and
would also match a logical progression (thus annotation collection
before creating an expressionset before plotting an expressionset before
differential expression before ontology searching).  That was wrong and
dumb.  Just use the collate field.  Unfortunately, that does not solve
my actual problem either, because the examples are actually run in
lexical order as they are found in the text files in man/.  I think
therefore I will need to use the @example tag in my roxygen (as opposed
to @examples (note the 's')) as per:

https://mirai-solutions.ch/techguides/advanced-usage-and-consistent-documentation.html

Section 1.3.1