random changes all around, here are the highlights:
1.  SVs now get incorporated into expts when running sva
2.  Made initial goseq2enrich and topgo2enrich[1]
3.  Small improvements to annotations and circos etc

[1]: I want to strip out all my ontology plotting functions and instead
use the ones provided by DOSE/clusterProfiler/enrichplot/etc.  To that
end, I want to have a function to convert the outputs from
goseq/topgo/gostats/gprofiler to clusterProfiler's enrichResult class.
In order to do this well, some significant changes will need to be made
to goseq and friends to simplify the datastructures they create.  Of
them, topgo's are the most problematic.