added a parameter when saving combined de results
Previously, if one saved the result from combine_de_results to a rda
file, the resulting file also included the entire result from
all_pairwise().  This is a bit wasteful of disk space; so by default it
no longer keeps that around.  The actual return of combine_de_tables()
still keeps that in the return list; but since R lists are able to just
pull from the set of bound object, it does not waste memory.