fixed a bug in combine_de_tables() uncovered by Zhezhen
One cannot guarantee a fixed table order across the various methods.  In
addition, if one explicitly selects a single de method, it assumes an
order which is effectively random (which I need to
investigate/understand).  I previously took a single index for each
table and assumed that would be constant across methods (which it
always has been, but I by default ask for every method); I moved the
scope of that index query into the block containing each individual
method.  This seems to have been sufficient to fix that problem, but it
does not explain why that order was not constant.