Mostly changes related to our newfound fascination with regression
I have a separate branch for the new model work, hopefully I didn't
pollute this with any of that, it is in a separate tree, so it should be
fine.  Instead, this should comprise some small functions to make the
various regression and model testing tasks we have tried easier.